Don't wait for the tip in send transaction tests, and try to speed up full lightwalletd syncs
