+++ zebrad/tests/common/sync.rs
@@ -45,6 +45,12 @@ pub const STOP_AT_HEIGHT_REGEX: &str = "stopping at configured height";
 pub const SYNC_FINISHED_REGEX: &str =
     r"finished initial sync to chain tip, using gossiped blocks .*sync_percent.*=.*100\.";
 
+/// The text that should be logged every time Zebra checks the sync progress.
+//
+// This is only used with `--feature lightwalletd-grpc-tests`
+#[allow(dead_code)]
+pub const SYNC_PROGRESS_REGEX: &str = r"sync_percent";
+
 /// The maximum amount of time Zebra should take to reload after shutting down.
 ///
 /// This should only take a second, but sometimes CI VMs or RocksDB can be slow.
